+++ cmd/gui/qtwidgets/escrow.go
@@ -131,7 +131,7 @@ func ShowEscrowPage() *widgets.QTabWidget {
 			ChainID:     ChainID,
 			Sender:      MainWallet.MainAddress,
 			SendingTime: common.GetCurrentTimeStampInSecond(),
-			Nonce:       int16(rand.Intn(0xffff)),
+			Nonce:       int64(rand.Intn(0xffff)),
 		}
 		tx := transactionsDefinition.Transaction{
 			TxData:    txd,
@@ -283,7 +283,7 @@ func ShowEscrowPage() *widgets.QTabWidget {
 	//		ChainID:     ChainID,
 	//		Sender:      MainWallet.MainAddress,
 	//		SendingTime: common.GetCurrentTimeStampInSecond(),
-	//		Nonce:       int16(rand.Intn(0xffff)),
+	//		Nonce:       int64(rand.Intn(0xffff)),
 	//	}
 	//	tx := transactionsDefinition.Transaction{
 	//		TxData:    txd,
