+++ cmd/website/handlers/transaction_handlers.go
@@ -288,7 +288,7 @@ func SendTransaction(w http.ResponseWriter, r *http.Request) {
 		ChainID:     int16(23),
 		Sender:      wl.MainAddress,
 		SendingTime: common.GetCurrentTimeStampInSecond(),
-		Nonce:       int16(rand.Intn(0xffff)),
+		Nonce:       int64(rand.Intn(0xffff)),
 	}
 
 	tx := transactionsDefinition.Transaction{
